+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Future Release
+
+### Changes
+
+- allow to pass function of arity one to `func` validation
+
 ## [1.2.4] - 2019.03.28
 
 ### Changes